Template releng ticket for upstream binary signing
The releng signing tickets for Butane and Ignition are currently created
by manually finding an old ticket, modifying it, copying the embedded
shell script to a file, running it, and seeing whether it works.  To avoid
all these manual steps, add a templated shell script that generates the
ticket contents, and that also knows how to directly extract and run the
embedded shell script for testing.