tmpl8: allow diffing against branch in pending PR
We've been diffing against the downstream main branch, which creates an
incentive to merge upstream template PRs expeditiously to keep the diff
clean.  However, it also adds noise to the downstream Git history.  It'd
be better to let changes accumulate in a PR, and merge it just before
starting a release.

Add tmpl8 command-line options allowing it to sync the cache (and thus
generate the diff) from the downstream PR branch instead.  If that branch
is missing, fall back to the downstream main branch as before.  This is
implemented via a regex and replacement string for repo URLs and a single
branch name for all repos, since that's all we actually need right now.

Hardcode arguments in the Makefile to match the values used by the GitHub
workflow.