Add GitHub workflow template for building and pushing containers
We're not onboarding every container-building workflow here, just the ones
that already follow a common template.

Explicitly set the arches parameter even when we're using the
actions-lib/build-container default.  Since we're templating the configs,
we can be a bit more verbose, and the workflow files are clearer when we
don't depend on a default from actions-lib.