Add GitHub workflow template for rebuilding release container
If there's a security issue in a container base image, we need to be able
to rebuild the container and update the registry tag that users are using.
For repos without Git tags, there's just `latest`/`main` and we can update
it by pushing a commit.  But for repos with a `release` tag aliased to
a versioned tag built from a Git tag, we need a way to create a new
registry tag and point `release` at it without having to tag a new
upstream release in Git.

Add a new workflow that takes an existing Git tag vX.Y.Z and a new
registry tag vX.Y.Z-A as input, builds the container, pushes it to the new
tag, and aliases `release` to it.  Enable it in the three upstream repos
that might potentially need it.