go/release-checklist: drop Butane MacPorts package
The package still exists, but I've dropped maintainership.  The FCOS docs
mention that the package is available in MacPorts, but I don't think it
needs to be kept 100% current or maintained directly by the CoreOS WG.
The package is marked openmaintainer, so anyone can submit an update PR
when they want.