rust: test vendor tarball before releasing
We've had two cases of cargo-vendor-filterer creating incorrect tarballs:

https://github.com/coreos/cargo-vendor-filterer/issues/27
https://github.com/coreos/cargo-vendor-filterer/issues/71

With the current checklist flow, we don't notice until the upstream
release is complete and we're doing a RHEL build with the vendored
tarball, forcing us to make a second release with a fixed tarball.
Catch this before release by generating a provisional vendor directory
during the prep phase and running `cargo test` against it.