tmpl8: fix cache sync for repos that haven't been forked
When syncing a cached Git repo, we try to pull from our fork's PR branch
when possible.  This lets us generate diffs that exclude any changes
already merged upstream, PRed downstream, but not yet landed downstream.

However, when onboarding a new Git repo, our fork is created by sync.yml
after the upstream onboarding PR is merged.  Thus, if the fork is missing,
the cache needs to fall back to the downstream default branch so upstream
PR CI can pass.  There's existing code to do that, but it didn't account
for the fact that GitHub pretends missing repos are private and asks for
authentication.  Therefore we'd block at a password prompt:

    Username for 'https://github.com':

Disable password prompting during this fetch so the fallback can run.

Fixes: https://github.com/coreos/repo-templates/issues/61
Fixes: d4f92803d3 ("tmpl8: allow diffing against branch in pending PR")