release-checklist: wait until mirror.openshift.com is updated
We had an instance recently where we thought the release was done, but
the ART ticket was blocked on more information and we didn't notice until
much later when users started wondering where the new version was.

Let's hard require the ART ticket to be closed before we can close the
release ticket.